Update ROS handling + Rename custom Node class + Move RosNode implementation to separate file
