+++ server/app.py
@@ -1,39 +1,14 @@
+# ROS
 import rclpy
 import signal
-from rclpy.node import Node
-from std_msgs.msg import String
 import threading
+# Custom ROS class
+from ros_handling import RosNode, ros2_thread
+# Flask
 from flask import Flask, send_from_directory, send_file
 # Path handling
 import os
 from pathlib import Path
-
-
-class TestPublisher(Node):
-    def __init__(self):
-        super().__init__('test_publisher')
-        self.publisher = self.create_publisher(String, 'flask_pub_topic', 10)
-        self.subscription = self.create_subscription(
-            String,
-            '/topic',
-            self.chatter_callback,
-            10)
-        self.latest_message = None
-
-    def chatter_callback(self, msg):
-        print(f'chatter cb received: {msg.data}')
-        self.latest_message = msg.data
-
-    def publish_message(self):
-        msg = String()
-        msg.data = 'hello, world!'
-        self.publisher.publish(msg)
-
-
-def ros2_thread(node):
-    print('entering ros2 thread')
-    rclpy.spin(node)
-    print('leaving ros2 thread')
 
 
 def sigint_handler(signal, frame):
@@ -52,10 +27,10 @@ def sigint_handler(signal, frame):
 
 
 rclpy.init(args=None)
-ros2_node = TestPublisher()
+node = RosNode()
 # Static folder is not equivalent to the static react build folder. React stores static build in 
 app = Flask(__name__, static_folder=Path(f"{os.getcwd()}/../react-app/build/"), static_url_path='/')
-threading.Thread(target=ros2_thread, args=[ros2_node]).start()
+threading.Thread(target=ros2_thread, args=[node]).start()
 prev_sigint_handler = signal.signal(signal.SIGINT, sigint_handler)
 
 # TODO
@@ -69,9 +44,9 @@ def serve_page():
 
 @app.route('/latest_message')
 def get_current_time():
-    return {'message': ros2_node.latest_message}
+    return {'message': node.latest_message}
 
 @app.route('/publish_message')
 def get_publish_message():
-    ros2_node.publish_message()
+    node.publish_message()
     return {}
